chore: Move legacy Frame logic into separate file
